+++ tests/testthat/test-get_pvol_de.R
@@ -1,5 +1,5 @@
 test_that("Pvol for German can be downloaded", {
-  skip_if_offline()
+  skip_if_offline(host = "opendata.dwd.de")
   time <- lubridate::floor_date(
     as.POSIXct(Sys.time(), tz = "Europe/Helsinki") - lubridate::hours(10),
     "5 mins"
@@ -9,5 +9,12 @@ test_that("Pvol for German can be downloaded", {
   expect_identical(
     lubridate::floor_date(pvol$datetime, "5 mins"),
     lubridate::with_tz(time, "UTC")
+  )
+})
+test_that("Correct error for old German data", {
+  skip_if_offline(host = "opendata.dwd.de")
+  expect_error(
+    get_pvol("deess", Sys.time() - lubridate::days(4)),
+    class = "getRad_error_germany_unexpected_number_of_files"
   )
 })